+++ frontend/spec/spec_helper.rb
@@ -30,7 +30,6 @@ require 'ffaker'
 require 'database_cleaner'
 
 require 'spree/testing_support/i18n' if ENV['CHECK_TRANSLATIONS']
-
 require 'spree/testing_support/authorization_helpers'
 require 'spree/testing_support/capybara_ext'
 require 'spree/testing_support/factories'
@@ -42,6 +41,7 @@ require 'spree/testing_support/order_walkthrough'
 require 'spree/testing_support/caching'
 require 'spree/testing_support/capybara_config'
 require 'spree/testing_support/image_helpers'
+require 'spree/testing_support/locale_helpers'
 require 'webdrivers'
 
 # Requires supporting files with custom matchers and macros, etc,
@@ -114,6 +114,7 @@ RSpec.configure do |config|
   config.include Spree::TestingSupport::ControllerRequests, type: :controller
   config.include Spree::TestingSupport::Flash
   config.include Spree::TestingSupport::ImageHelpers
+  config.include Spree::TestingSupport::LocaleHelpers
 
   config.order = :random
   Kernel.srand config.seed
